Remove unnecessary utility method
